fix(graph): keep Kuzu wiring cgo-safe
