+++ internal/config/config_test.go
@@ -124,6 +124,7 @@ func TestLoadRejectsMissingPostgresDSN(t *testing.T) {
 }
 
 func TestLoadRejectsMissingKuzuPath(t *testing.T) {
+	t.Setenv("CEREBRO_KUZU_PATH", "")
 	t.Setenv("CEREBRO_GRAPH_STORE_DRIVER", GraphStoreDriverKuzu)
 	if _, err := Load(); err == nil {
 		t.Fatal("Load() error = nil, want non-nil")
