+++ internal/graphstore/kuzu/kuzu.go
@@ -1,3 +1,5 @@
+//go:build cgo
+
 package kuzu
 
 import (
