+++ internal/graphstore/kuzu/kuzu_test.go
@@ -1,3 +1,5 @@
+//go:build cgo
+
 package kuzu
 
 import (
